Fix Calendar component. Pass selected days as props
