+++ src/features/event/components/Calendar/Calendar.tsx
@@ -1,10 +1,9 @@
-import { useState, useEffect } from 'react';
-import { parseISO } from 'date-fns';
-import { useEventDateStore } from '@/stores/eventDateStore';
 import { Calendar as CalendarUi, type CalendarProps } from '@/components/ui/calendar';
 import { CalendarFooter } from './CalendarFooter';
 
-type CalendarPropsType = Omit<
+type CalendarPropsType = {
+  selectedDays: Date[] | undefined
+} & Omit<
   CalendarProps,
   | 'selected'
   | 'mode'
@@ -12,21 +11,11 @@ type CalendarPropsType = Omit<
 >
 
 export const Calendar = (props: CalendarPropsType) => {
-  const { eventDates } = useEventDateStore();
-  // State for selected days in the calendar
-  const [calendarSelectedDays, setCalendarSelectedDays] = useState<Date[] | undefined>();
-
-  useEffect(() => {
-    // Update calendar selection status based on the store
-    const updatedSelectedDays = eventDates.map(eventDate => parseISO(eventDate.date));
-    setCalendarSelectedDays(updatedSelectedDays);
-  }, [eventDates]);
-
   return (
     <CalendarUi
       mode='multiple'
-      selected={calendarSelectedDays}
-      footer={<CalendarFooter daysCount={calendarSelectedDays?.length || 0} />}
+      selected={props.selectedDays}
+      footer={<CalendarFooter daysCount={props?.selectedDays?.length || 0} />}
       {...props}
     />
   );
